Extract DefaultAnalysisContext
