+++ analyzers/src/SonarAnalyzer.CSharp/Rules/SymbolicExecution/DefaultAnalysisContext.cs
@@ -1,0 +1,45 @@
+﻿/*
+ * SonarAnalyzer for .NET
+ * Copyright (C) 2015-2020 SonarSource SA
+ * mailto: contact AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.CodeAnalysis;
+
+namespace SonarAnalyzer.Rules.SymbolicExecution
+{
+    internal abstract class DefaultAnalysisContext : ISymbolicExecutionAnalysisContext
+    {
+        private readonly List<Location> locations = new List<Location>();
+
+        public bool SupportsPartialResults => false;
+
+        public IEnumerable<Diagnostic> GetDiagnostics() =>
+            locations.Distinct().Select(CreateDiagnostic);
+
+        public void AddLocation(Location location) => locations.Add(location);
+
+        public void Dispose()
+        {
+            // Nothing to do here.
+        }
+
+        protected abstract Diagnostic CreateDiagnostic(Location location);
+    }
+}
